requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
mlflow
apache-airflow
joblib
statsmodels
arch
shap